Move accountFromAny into new module and create a test
