+++ packages/stargate/src/stargateclient.ts
@@ -5,30 +5,20 @@ import {
   isSearchByHeightQuery,
   isSearchBySentFromOrToQuery,
   isSearchByTagsQuery,
-  PubKey,
   SearchTxFilter,
   SearchTxQuery,
 } from "@cosmjs/launchpad";
-import { Uint53, Uint64 } from "@cosmjs/math";
-import { decodePubkey } from "@cosmjs/proto-signing";
+import { Uint53 } from "@cosmjs/math";
 import {
   broadcastTxCommitSuccess,
   Tendermint34Client,
   toRfc3339WithNanoseconds,
 } from "@cosmjs/tendermint-rpc";
-import { assert, assertDefinedAndNotNull } from "@cosmjs/utils";
-import Long from "long";
+import { assertDefinedAndNotNull } from "@cosmjs/utils";
 
-import { BaseAccount, ModuleAccount } from "./codec/cosmos/auth/v1beta1/auth";
+import { Account, accountFromAny } from "./accounts";
 import { MsgData, TxMsgData } from "./codec/cosmos/base/abci/v1beta1/abci";
 import { Coin } from "./codec/cosmos/base/v1beta1/coin";
-import {
-  BaseVestingAccount,
-  ContinuousVestingAccount,
-  DelayedVestingAccount,
-  PeriodicVestingAccount,
-} from "./codec/cosmos/vesting/v1beta1/vesting";
-import { Any } from "./codec/google/protobuf/any";
 import { AuthExtension, BankExtension, QueryClient, setupAuthExtension, setupBankExtension } from "./queries";
 
 /** A transaction that is indexed as part of the transaction history */
@@ -40,14 +30,6 @@ export interface IndexedTx {
   readonly code: number;
   readonly rawLog: string;
   readonly tx: Uint8Array;
-}
-
-export interface Account {
-  /** Bech32 account address */
-  readonly address: string;
-  readonly pubkey: PubKey | null;
-  readonly accountNumber: number;
-  readonly sequence: number;
 }
 
 export interface SequenceResponse {
@@ -90,70 +72,6 @@ export function assertIsBroadcastTxSuccess(
     throw new Error(
       `Error when broadcasting tx ${result.transactionHash} at height ${result.height}. Code: ${result.code}; Raw log: ${result.rawLog}`,
     );
-  }
-}
-
-function uint64FromProto(input: number | Long | null | undefined): Uint64 {
-  if (!input) return Uint64.fromNumber(0);
-  return Uint64.fromString(input.toString());
-}
-
-function accountFromBaseAccount(input: BaseAccount): Account {
-  const { address, pubKey, accountNumber, sequence } = input;
-  const pubkey = decodePubkey(pubKey);
-  return {
-    address: address,
-    pubkey: pubkey,
-    accountNumber: uint64FromProto(accountNumber).toNumber(),
-    sequence: uint64FromProto(sequence).toNumber(),
-  };
-}
-
-/**
- * Takes an `Any` encoded account from the chain and extracts some common
- * `Account` information from it. This is supposed to support the most relevant
- * common Cosmos SDK account types. If you need support for exotix account types,
- * you'll need to write your own account decoder.
- */
-export function accountFromAny(input: Any): Account {
-  const { typeUrl, value } = input;
-
-  switch (typeUrl) {
-    // auth
-
-    case "/cosmos.auth.v1beta1.BaseAccount":
-      return accountFromBaseAccount(BaseAccount.decode(value));
-    case "/cosmos.auth.v1beta1.ModuleAccount": {
-      const baseAccount = ModuleAccount.decode(value).baseAccount;
-      assert(baseAccount);
-      return accountFromBaseAccount(baseAccount);
-    }
-
-    // vesting
-
-    case "/cosmos.vesting.v1beta1.BaseVestingAccount": {
-      const baseAccount = BaseVestingAccount.decode(value)?.baseAccount;
-      assert(baseAccount);
-      return accountFromBaseAccount(baseAccount);
-    }
-    case "/cosmos.vesting.v1beta1.ContinuousVestingAccount": {
-      const baseAccount = ContinuousVestingAccount.decode(value)?.baseVestingAccount?.baseAccount;
-      assert(baseAccount);
-      return accountFromBaseAccount(baseAccount);
-    }
-    case "/cosmos.vesting.v1beta1.DelayedVestingAccount": {
-      const baseAccount = DelayedVestingAccount.decode(value)?.baseVestingAccount?.baseAccount;
-      assert(baseAccount);
-      return accountFromBaseAccount(baseAccount);
-    }
-    case "/cosmos.vesting.v1beta1.PeriodicVestingAccount": {
-      const baseAccount = PeriodicVestingAccount.decode(value)?.baseVestingAccount?.baseAccount;
-      assert(baseAccount);
-      return accountFromBaseAccount(baseAccount);
-    }
-
-    default:
-      throw new Error(`Unsupported type: '${typeUrl}'`);
   }
 }
